JACOBIN-780 undo junie choice to put function bigdecimalDivideMathContext in a separate file
